Restructuring the command tree to be on the "noun verb" form instead of "verb noun"
