+++ cmd/analyze.go
@@ -2,10 +2,13 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/pkg/browser"
 	"github.com/qlik-oss/corectl/internal"
 	"github.com/qlik-oss/corectl/printer"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 	"os"
+	"strings"
 )
 
 var getAssociationsCmd = &cobra.Command{
@@ -93,3 +96,44 @@ var getKeysCmd = &cobra.Command{
 		printer.PrintFields(data, true)
 	},
 }
+
+var evalCmd = &cobra.Command{
+	Use:   "eval <measure 1> [<measure 2...>] by <dimension 1> [<dimension 2...]",
+	Short: "Evaluates a list of measures and dimensions",
+	Long:  `Evaluates a list of measures and dimensions. To evaluate a measure for a specific dimension use the <measure> by <dimension> notation. If dimensions are omitted then the eval will be evaluated over all dimensions.`,
+	Example: `corectl eval "Count(a)" // returns the number of values in field "a"
+corectl eval "1+1" // returns the calculated value for 1+1
+corectl eval "Avg(Sales)" by "Region" // returns the average of measure "Sales" for dimension "Region"
+corectl eval by "Region" // Returns the values for dimension "Region"`,
+
+	Run: func(ccmd *cobra.Command, args []string) {
+		if len(args) == 0 {
+			fmt.Println("Expected at least one dimension or measure")
+			ccmd.Usage()
+			os.Exit(1)
+		}
+		state := internal.PrepareEngineState(rootCtx, headers, false)
+		internal.Eval(rootCtx, state.Doc, args)
+	},
+}
+
+var catwalkCmd = withCommonLocalFlags(&cobra.Command{
+	Use:   "catwalk",
+	Short: "Opens the specified app in catwalk",
+	Long:  `Opens the specified app in catwalk. If no app is specified the catwalk hub will be opened.`,
+	Example: `corectl catwalk --app my-app.qvf
+corectl catwalk --app my-app.qvf --catwalk-url http://localhost:8080`,
+
+	Run: func(ccmd *cobra.Command, args []string) {
+		catwalkURL := viper.GetString("catwalk-url") + "?engine_url=" + internal.TidyUpEngineURL(viper.GetString("engine")) + "/apps/" + viper.GetString("app")
+		if !strings.HasPrefix(catwalkURL, "www") && !strings.HasPrefix(catwalkURL, "https://") && !strings.HasPrefix(catwalkURL, "http://") {
+			fmt.Println("Please provide a valid URL starting with 'https://', 'http://' or 'www'")
+			os.Exit(1)
+		}
+		err := browser.OpenURL(catwalkURL)
+		if err != nil {
+			fmt.Println("Could not open URL", err)
+			os.Exit(1)
+		}
+	},
+}, "catwalk-url")
