+++ test/golden/help-2.golden
@@ -4,16 +4,18 @@ Usage:
   corectl [flags]
   corectl [command]
 
-Basic Commands:
-  all           Sets the objects, measures, dimensions, connections and script in the current app
-  assoc         Print table associations summary
+App Building Commands:
   build         Reloads and saves the app after updating connections, dimensions, measures, objects and the script
+  reload        Reloads the app.
+
+App Analysis Commands:
+  assoc         Print table associations summary
+  catwalk       Opens the specified app in catwalk
   eval          Evaluates a list of measures and dimensions
   field         Shows content of a field
   fields        Print field list
   keys          Print key-only field list
   meta          Shows metadata about the app
-  reload        Reloads the app.
   tables        Print tables summary
 
 Advanced Commands:
@@ -25,7 +27,6 @@ Advanced Commands:
   script        Explore and manage the script
 
 Other Commands:
-  catwalk       Opens the specified app in catwalk
   completion    Generates auto completion scripts
   help          Help about any command
   status        Prints status info about the connection to the engine and current app
@@ -38,8 +39,6 @@ Flags:
       --headers stringToString   Http headers to use when connecting to Qlik Associative Engine (default [])
   -h, --help                     help for corectl
       --no-data                  Open app without data
-      --no-save                  Do not save the app
-      --suppress                 Suppress all confirmation dialogues
   -t, --traffic                  Log JSON websocket traffic to stdout
       --ttl string               Qlik Associative Engine session time to live in seconds (default "30")
   -v, --verbose                  Logs extra information
