+++ test/golden/project1-script-2.golden
@@ -1,25 +1,2 @@
-Explore and manage the script
+Saving app... Done
 
-Usage:
-  corectl script [command]
-
-Available Commands:
-  script      Sets the script in the current app
-  script      Print the reload script
-
-Flags:
-  -h, --help   help for script
-
-Global Flags:
-  -a, --app string               App name, if no app is specified a session app is used instead.
-  -c, --config string            path/to/config.yml where parameters can be set instead of on the command line
-  -e, --engine string            URL to the Qlik Associative Engine (default "localhost:9076")
-      --headers stringToString   Http headers to use when connecting to Qlik Associative Engine (default [])
-      --no-data                  Open app without data
-      --no-save                  Do not save the app
-      --suppress                 Suppress all confirmation dialogues
-  -t, --traffic                  Log JSON websocket traffic to stdout
-      --ttl string               Qlik Associative Engine session time to live in seconds (default "30")
-  -v, --verbose                  Logs extra information
-
-Use "corectl script [command] --help" for more information about a command.
